ExamOfUser entity and its CRUD, and GetUsersExamResultInfo method Added

+++ Business/Rules/BusinessRules/CourseBusinessRules.cs
@@ -23,7 +23,7 @@ namespace Business.Rules.BusinessRules
             _examService = examService;
         }
 
-        public async Task ValidateExamPoint(int examPoint)
+        public async Task ValidateExamPoint(float examPoint)
         {
             if (examPoint < 0 || examPoint > 100)
             {
